+++ core/Config.php
@@ -179,6 +179,18 @@ class Config
     }
 
     /**
+     * Run a callable functions for the defined env when toArray() is called.
+     *
+     * @param callable $fn The function to run, the first argument of the closure is the {{luya\Config}} object.
+     * @return ConfigDefinition
+     * @since 1.0.23
+     */
+    public function callback(callable $fn)
+    {
+        return $this->addDefinition(New ConfigDefinition(ConfigDefinition::GROUP_CALLABLE, false, $fn));
+    }
+
+    /**
      * Register a component
      *
      * @param string $id The id of the component
@@ -218,7 +230,7 @@ class Config
     private $_definitions = [];
 
     /**
-     * Add a defintion into the defintions bag.
+     * Add a definition into the definitions bag.
      *
      * @param ConfigDefinition $definition
      * @return ConfigDefinition
@@ -291,7 +303,7 @@ class Config
     }
 
     /**
-     * Append a given defintion int othe config
+     * Append a given definition int othe config
      *
      * @param array $config
      * @param ConfigDefinition $definition
@@ -319,11 +331,16 @@ class Config
                 foreach ($definition->getConfig() as $v) {
                     $config['bootstrap'][] = $v;
                 }
+                break;
+
+            case ConfigDefinition::GROUP_CALLABLE:
+                call_user_func($definition->getConfig(), $this);
+            break;
         }
     }
 
     /**
-     * Add a array key based component defintion.
+     * Add a array key based component definition.
      *
      * @param array $config
      * @param ConfigDefinition $definition